requirements.txt
pandas==1.0.4
biopython==1.77
click==7.1.1
Flask==1.1.2
Jinja2==2.11.2
mysql-connector-python==8.0.20
numpy==1.18.5



